+++ Editor/Drawers/PropRangeDrawer/PropRangeAttributeDrawer.cs
@@ -113,6 +113,26 @@ namespace SaintsField.Editor.Drawers.PropRangeDrawer
             return AdaptAttributeDrawer.GetDoubleValuePost(value);
         }
 
+        private static float GetValue(MetaInfo metaInfo, float newValue)
+        {
+            // property.floatValue = newValue;
+            float step = metaInfo.Step;
+            bool isFloat = metaInfo.IsFloat;
+            // Debug.Log(step);
+            if (step <= 0)
+            {
+                // return newValue;
+                return Mathf.Clamp(newValue, metaInfo.MinValue, metaInfo.MaxValue);
+            }
+
+            if (isFloat)
+            {
+                return Util.BoundFloatStep(newValue, metaInfo.MinValue, metaInfo.MaxValue, step);
+            }
+
+            return Util.BoundIntStep(newValue, metaInfo.MinValue, metaInfo.MaxValue, (int)step);
+        }
+
         public AutoRunnerFixerResult AutoRunFix(PropertyAttribute propertyAttribute, IReadOnlyList<PropertyAttribute> allAttributes,
             SerializedProperty property, MemberInfo memberInfo, object parent)
         {