+++ Samples~/Scripts/PlayaArraySizeExample.cs
@@ -1,10 +1,11 @@
 using System;
 using SaintsField.Playa;
 using SaintsField.Samples.Scripts.SaintsEditor;
+using UnityEngine;
 
 namespace SaintsField.Samples.Scripts
 {
-    public class PlayaArraySizeExample : SaintsMonoBehaviour
+    public class PlayaArraySizeExample : MonoBehaviour
     {
         [ArraySize(3)] public int[] myArr3;