+++ Samples~/Scripts/SaintsEditor/Testing/ListFallFlow.cs
@@ -20,8 +20,16 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
 
         // [SaintsRow]
         public MyStruct[] structArr;
+        [ListDrawerSettings(searchable: true)]
+        public MyStruct[] structArrLis;
 
+        [LayoutStart("Horizontal", ELayout.Horizontal)]
         [PropRange(0, 1)]
         public float[] floatArr;
+
+        [RichLabel("$" + nameof(ItemValue))]
+        public string[] g23;
+
+        private string ItemValue(string v) => v;
     }
 }